requirements.txt
# Core LangChain and Anthropic dependencies
langchain-anthropic>=0.3.18
langchain-core>=0.3.0
langchain-community>=0.3.0
anthropic>=0.40.0  # Direct Anthropic API client for model listing

# Environment and configuration
python-dotenv>=1.0.0

# Terminal UI enhancements
rich>=13.0.0
colorama>=0.4.6
textual>=5.2.0

# Optional: For advanced features
tiktoken>=0.5.0  # Token counting
